initialize applugin module with __init__.py
